refactor: 렌더링 폴더명 변경
