+++ package.json
@@ -4,13 +4,15 @@
   "description": "뉴스스탠드 기본화면",
   "main": "src/js/crawler/Crawler.js",
   "scripts": {
-    "start": "node src/js/crawler/Crawler.js"
+    "start": "node src/js/crawler/Crawler.js",
+    "watch": "json-server --watch src/data/subscribedNews.json"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "type": "module",
   "dependencies": {
+    "json-server": "^1.0.0-alpha.23",
     "puppeteer": "^22.4.1"
   }
 }
