+++ src/js/view/NewsStandRenderer.js
