+++ src/js/view/NewsTitleRenderer.js
