+++ src/js/view/Utils.js
